+++ test_case/adobe_xd/lib/XDColors.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class XDColors {
+  static const Color theme = const Color(0xff254f6e);
+  static const Color highlight = const Color(0xff00a9de);
+  static const Color white = const Color(0xffffffff);
+  static const Color grey0 = const Color(0xffd5d5d5);
+  static const Color grey1 = const Color(0xff808080);
+  static const Color grey2 = const Color(0xff333333);
+
+  static const List<Color> grey = const [grey0, grey1, grey2];
+}