+++ test_case/best_flutter_ui_templates/lib/navigation_home_screen.dart
@@ -1,0 +1,84 @@
+import 'package:best_flutter_ui_templates/app_theme.dart';
+import 'package:best_flutter_ui_templates/custom_drawer/drawer_user_controller.dart';
+import 'package:best_flutter_ui_templates/custom_drawer/home_drawer.dart';
+import 'package:best_flutter_ui_templates/home_screen.dart';
+import 'package:fair/fair.dart';
+import 'package:flutter/material.dart';
+
+class NavigationHomeScreen extends StatefulWidget {
+  @override
+  _NavigationHomeScreenState createState() => _NavigationHomeScreenState();
+}
+
+class _NavigationHomeScreenState extends State<NavigationHomeScreen> {
+  Widget screenView;
+  DrawerIndex drawerIndex;
+
+  @override
+  void initState() {
+    drawerIndex = DrawerIndex.HOME;
+    screenView = const MyHomePage();
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: AppTheme.nearlyWhite,
+      child: SafeArea(
+        top: false,
+        bottom: false,
+        child: Scaffold(
+          backgroundColor: AppTheme.nearlyWhite,
+          body: DrawerUserController(
+            screenIndex: drawerIndex,
+            drawerWidth: MediaQuery.of(context).size.width * 0.75,
+            onDrawerCall: (DrawerIndex drawerIndexdata) {
+              changeIndex(drawerIndexdata);
+              //callback from drawer for replace screen as user need with passing DrawerIndex(Enum index)
+            },
+            screenView: screenView,
+            //we replace screen view as we need on navigate starting screens like MyHomePage, HelpScreen, FeedbackScreen, etc...
+          ),
+        ),
+      ),
+    );
+  }
+
+  void changeIndex(DrawerIndex drawerIndexdata) {
+    if (drawerIndex != drawerIndexdata) {
+      drawerIndex = drawerIndexdata;
+      if (drawerIndex == DrawerIndex.HOME) {
+        setState(() {
+          screenView = const MyHomePage();
+        });
+      } else if (drawerIndex == DrawerIndex.Help) {
+        setState(() {
+          // screenView = HelpScreen();
+          screenView = FairWidget(
+            name: 'help',
+            path: 'assets/bundle/lib_help_screen.fair.bin',
+          );
+        });
+      } else if (drawerIndex == DrawerIndex.FeedBack) {
+        setState(() {
+          // screenView = FeedbackScreen();
+          screenView = FairWidget(
+            name: 'feedback',
+            path: 'assets/bundle/lib_feedback_screen.fair.bin',
+          );
+        });
+      } else if (drawerIndex == DrawerIndex.Invite) {
+        setState(() {
+          // screenView = InviteFriend();
+          screenView = FairWidget(
+            name: 'invite',
+            path: 'assets/bundle/lib_invite_friend_screen.fair.bin',
+          );
+        });
+      } else {
+        //do in your way......
+      }
+    }
+  }
+}