+++ test_case/fair_mobx/android/app/src/main/kotlin/com/mobxwithfair/fairmobx/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.mobxwithfair.fairmobx
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}