+++ test_case/fair_provider/android/app/src/main/kotlin/com/providerwithfair/demo/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.providerwithfair.demo
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}