+++ test_case/flutter_component/lib/src/generated.fair.dart
@@ -1642,15 +1642,15 @@ class AppGeneratedModule extends GeneratedModule {
       },
       'RotateAnimatedTextKit': (props) => RotateAnimatedTextKit(
             key: props['key'],
-            text: (props['text']),//as<String>(props['text'])
+            text: (props['text']).cast<String>(),//as<String>(props['text'])
             textStyle: props['textStyle'],
             transitionHeight: props['transitionHeight']?.toDouble(),
-            pause: props['pause'],
+            pause: props['pause'] ?? const Duration(milliseconds: 500),
             onNext: props['onNext'],
             onNextBeforePause: props['onNextBeforePause'],
             onFinished: props['onFinished'],
             totalRepeatCount: props['totalRepeatCount'] ?? 3,
-            duration: props['duration'],
+            duration: props['duration'] ?? const Duration(milliseconds: 2000),
             onTap: props['onTap'],
             alignment: props['alignment'] ?? const Alignment(0.0, 0.0),
             textAlign: props['textAlign'] ?? TextAlign.start,
@@ -1670,7 +1670,7 @@ class AppGeneratedModule extends GeneratedModule {
                 props['animationDuration'] ?? const Duration(milliseconds: 270),
             mainAxisAlignment:
                 props['mainAxisAlignment'] ?? MainAxisAlignment.spaceBetween,
-            items: (props['items']),//as<BottomNavyBarItem>(props['items']
+            items: (props['items']).cast<BottomNavyBarItem>(),//as<BottomNavyBarItem>(props['items']
             onItemSelected: props['onItemSelected'],
             curve: props['curve'] ?? Curves.linear,
           ),