+++ flutter_version/flutter_2_5_0/CHANGELOG.md
@@ -1,5 +1,13 @@
+## 2.5.0+1
+* Generated from flutter 2.5.0
+* Add the LinearGradient to the mapping table
+
 ## 2.5.0
 * Generated from flutter 2.5.0
+
+## 2.0.6+2
+* Generated from flutter 2.0.6
+* Add the LinearGradient to the mapping table
 
 ## 2.0.6+1
 * Generated from flutter 2.0.6